+++ R/ui-creationParamsTable.R
@@ -24,10 +24,10 @@ creationParamsTable <- function(rules) {
                   "()](https://beerda.github.io/nuggets/reference/",
                   aa$call_function,
                   ".html) with the following parameters:")
-    fun <- markdown(fun)
+    fun <- shiny::markdown(fun)
 
     args <- lapply(aa$call_args, function(x) {
-        markdown(paste0("```r\n",
+        shiny::markdown(paste0("```r\n",
                         paste(deparse(x), collapse = ""),
                         "\n```"))
     })
@@ -35,20 +35,20 @@ creationParamsTable <- function(rules) {
     # do not present arguments containing variable names as usual values
     for (x in c("x")) {
         if (!is.null(args[[x]])) {
-            args[[x]] <- markdown(paste0("```r\n",
+            args[[x]] <- shiny::markdown(paste0("```r\n",
                                          paste(aa$call_args[[x]], collapse = ""),
                                          "\n```"))
         }
     }
 
     htmlrows <- lapply(seq_along(args), function(i) {
-        tags$tr(tags$td(paste(names(args)[i], "=")),
-                tags$td(args[[i]]))
+        htmltools::tags$tr(htmltools::tags$td(paste(names(args)[i], "=")),
+                htmltools::tags$td(args[[i]]))
     })
 
-    tagList(fun,
-            br(),
-            do.call(tags$table,
+    shiny::tagList(fun,
+            htmltools::br(),
+            do.call(htmltools::tags$table,
                     c(list(class = "info-table left", width = "100%"),
                       htmlrows)))
 }