+++ R/ui-highlightCondition.R
@@ -19,7 +19,7 @@
 
 highlightCondition <- function(x) {
     x <- gsub("[{}]", "", x)
-    x <- htmlEscape(x)
+    x <- htmltools::htmlEscape(x)
     x <- gsub("=", "</span>=<span class=\"pred_v\">", x)
     x <- gsub("^", "<span class=\"pred_n\">", x)
     x <- gsub("$", "</span>", x)