+++ tests/testthat/test-ui-callDataTable.R
@@ -18,6 +18,8 @@
 
 
 test_that("callDataTable()", {
+    .skip_if_shiny_not_installed()
+
     meta <- data.frame(
         data_name = c("condA", "condB", "numX"),
         type = c("condition", "condition", "numeric"),