+++ NEWS.md
@@ -1,3 +1,8 @@
+# nuggets 2.2.0
+- released: ???
+- moved Shiny-related packages (shiny, shinyjs, shinyWidgets, DT, htmltools,
+  htmlwidgets, jsonlite) from Imports to Suggests
+
 # nuggets 2.1.0
 - released: 2025-11-05
 - added cluster_associations()